fix(aces): resolve PR #817's SonarCloud file-length and return-count findings
- aces_content_realization.py grew to 568 lines (SonarCloud S104, 500
  max) after the content-source policy fix. Splits into three
  modules along their existing natural seam (the file's own docstring
  already distinguished "content-placement" from "image-free
  content"): aces_content_source_policy.py (the deny-list/allowlist
  policy itself, reusable by both realization paths),
  aces_image_free_content_realization.py (the ADR-048 image-free
  entry point and its helpers), and aces_content_realization.py (the
  legacy Compose-managed path, now back under budget).
- _forbidden_source_reason (now forbidden_source_reason) and
  _project_source_rejection each grew to 4 return statements (S1142,
  3 max) from the new deny-list check. Consolidated each to a single
  trailing return over a computed reason/rejection value.
