+++ src/aptl/backends/aces_content_realization.py
@@ -38,6 +38,7 @@ from typing import Any
 from aces_contracts.diagnostics import Diagnostic
 from aces_contracts.planning import PlannedResource
 
+from aptl.backends.aces_content_source_policy import forbidden_source_reason
 from aptl.backends.aces_diagnostics import diagnostic
 from aptl.backends.aces_realization_values import (
     content_source_name as _content_source_name,
@@ -64,62 +65,6 @@ _CONTENT_REALIZABLE_SERVICES: dict[str, str] = {
 
 _RUNTIME_OBSERVED_PREFIX = "runtime-observed:"
 _SAFE_DEST_RELPATH = re.compile(r"^[A-Za-z0-9][A-Za-z0-9._/-]*$")
-
-# Pure secret material with no legitimate scenario-content use, regardless
-# of project-root containment (issue #816): a content placement naming one
-# of these as its source could copy a real operator credential into any
-# addressed node, including one a lab participant can reach (e.g. kali over
-# SSH). Checked before containment so nothing about the target node or the
-# author's self-reported `sensitive` flag can authorize it.
-_FORBIDDEN_SOURCE_PREFIXES = (
-    ".git/",
-    "config/soc_certs/",
-    "config/wazuh_indexer_ssl_certs/",
-)
-
-# keys/ and config/lab-ssh/ are the one exception: they hold both the
-# public/combined authorized_keys files legitimately distributed onto
-# target nodes and the SEC #417 pivot private keys intentionally placed
-# onto their designated node (src/aptl/core/ssh.py). A blanket deny would
-# break those already-shipped placements, so instead only the exact
-# filenames that module generates are allowed; anything else under these
-# directories is unexpected and rejected the same as a forbidden path.
-_KEY_MATERIAL_PREFIXES = ("keys/", "config/lab-ssh/")
-_ALLOWED_KEY_SOURCE_NAMES = frozenset(
-    {
-        "keys/aptl_lab_key.pub",
-        "keys/authorized_keys",
-        "keys/target_authorized_keys",
-        "keys/victim_authorized_keys",
-        "config/lab-ssh/kali_pivot_key",
-        "config/lab-ssh/kali_pivot_key.pub",
-        "config/lab-ssh/workstation_pivot_key",
-        "config/lab-ssh/workstation_pivot_key.pub",
-    }
-)
-
-
-def _forbidden_source_reason(source_name: str) -> str | None:
-    """Return a reason code if `source_name` must never be a content source.
-
-    Independent of, and checked before, the project-containment check: a
-    path can be entirely inside the project root and still be something a
-    scenario must never be able to select (issue #816).
-    """
-
-    normalized = source_name.lstrip("/")
-    if normalized == ".env" or normalized.startswith(".env."):
-        return "source-is-environment-file"
-    if normalized == ".git" or any(
-        normalized.startswith(prefix) for prefix in _FORBIDDEN_SOURCE_PREFIXES
-    ):
-        return "source-is-forbidden-path"
-    if (
-        any(normalized.startswith(prefix) for prefix in _KEY_MATERIAL_PREFIXES)
-        and normalized not in _ALLOWED_KEY_SOURCE_NAMES
-    ):
-        return "source-is-unlisted-key-material"
-    return None
 
 
 @dataclass(frozen=True)
@@ -376,7 +321,7 @@ def _resolve_project_source(
 ) -> tuple[Path | None, list[Diagnostic]]:
     """Resolve a checked-in source path, failing closed on containment escape."""
 
-    forbidden_reason = _forbidden_source_reason(source_name)
+    forbidden_reason = forbidden_source_reason(source_name)
     if forbidden_reason is not None:
         return None, [_reject(address, forbidden_reason)]
     try:
@@ -413,155 +358,3 @@ def _reject(address: str, reason_code: str) -> Diagnostic:
             f"realization policy (reason={reason_code})."
         ),
     )
-
-
-_ImageFreeResult = tuple[DeploymentContentRealization | None, list[Diagnostic]]
-
-
-def _inline_text_image_free_placement(
-    resource: PlannedResource,
-    target_address: str,
-    *,
-    dest: str,
-    name: str,
-    text: str | None,
-    spec: Mapping[str, Any],
-) -> _ImageFreeResult | None:
-    """Lower an inline-text placement, or None if this spec is not one."""
-
-    if text is None or not dest:
-        return None
-    return (
-        DeploymentContentRealization(
-            address=resource.address,
-            target_address=target_address,
-            content_name=name,
-            volume_suffix="",
-            dest_relpath=dest.lstrip("/"),
-            source_kind="inline-text",
-            inline_text=text,
-            sensitive=spec.get("sensitive") is True,
-        ),
-        [],
-    )
-
-
-def _project_source_rejection(
-    resource: PlannedResource, source_name: str
-) -> list[Diagnostic] | None:
-    """Return fail-closed diagnostics if a project source can't be realized, else None."""
-
-    if source_name.startswith(_RUNTIME_OBSERVED_PREFIX):
-        return [
-            diagnostic(
-                "aptl.provisioner.content-not-realizable",
-                resource.address,
-                "runtime-observed content cannot be recreated from source.",
-            )
-        ]
-    forbidden_reason = _forbidden_source_reason(source_name)
-    if forbidden_reason is not None:
-        return [
-            diagnostic(
-                "aptl.provisioner.content-source-forbidden",
-                resource.address,
-                f"content source is not permitted (reason={forbidden_reason}).",
-            )
-        ]
-    try:
-        _resolve_within_project(Path(""), source_name)
-    except PathContainmentError:
-        return [
-            diagnostic(
-                "aptl.provisioner.content-source-escapes-project",
-                resource.address,
-                "content source path escapes the project root.",
-            )
-        ]
-    return None
-
-
-def _project_source_image_free_placement(
-    resource: PlannedResource,
-    target_address: str,
-    *,
-    dest: str,
-    name: str,
-    source_name: str,
-    content_type: str,
-    spec: Mapping[str, Any],
-) -> _ImageFreeResult | None:
-    """Lower a project-contained source placement, or None if this spec is not one."""
-
-    if not source_name or not dest:
-        return None
-    diagnostics = _project_source_rejection(resource, source_name)
-    if diagnostics is not None:
-        return None, diagnostics
-    kind = "project-directory" if content_type == "directory" else "project-file"
-    return (
-        DeploymentContentRealization(
-            address=resource.address,
-            target_address=target_address,
-            content_name=name,
-            volume_suffix="",
-            dest_relpath=dest.lstrip("/"),
-            source_kind=kind,
-            source_relpath=source_name,
-            sensitive=spec.get("sensitive") is True,
-        ),
-        [],
-    )
-
-
-def resolve_image_free_content_placement(
-    resource: PlannedResource,
-    payload: Mapping[str, Any],
-    target_address: str,
-) -> _ImageFreeResult:
-    """Resolve content for an image-free node (ADR-048).
-
-    The generic materializer places declared config directly into the node's
-    container, so there is no compose service / named-volume requirement.
-    ``path``/``destination`` is the authored, literal absolute destination
-    (never volume-relative). Inline text and project-contained file/directory
-    sources both lower to a ``DeploymentContentRealization``; dataset content
-    and runtime-observed sources are not realizable and fail closed rather
-    than being silently dropped.
-    """
-
-    spec = _placement_spec(payload)
-    if spec is None:
-        return None, [
-            diagnostic(
-                "aptl.provisioner.invalid-content-spec",
-                resource.address,
-                "content placement has no spec.",
-            )
-        ]
-    dest = _optional_string(spec, "path") or _optional_string(spec, "destination")
-    name = _optional_string(payload, "content_name") or _optional_string(payload, "name") or ""
-
-    result = _inline_text_image_free_placement(
-        resource, target_address, dest=dest, name=name, text=_content_text(spec), spec=spec
-    )
-    if result is None:
-        result = _project_source_image_free_placement(
-            resource,
-            target_address,
-            dest=dest,
-            name=name,
-            source_name=_content_source_name(spec),
-            content_type=_optional_string(spec, "type") or "",
-            spec=spec,
-        )
-    if result is None:
-        result = None, [
-            diagnostic(
-                "aptl.provisioner.image-free-content-unsupported",
-                resource.address,
-                "image-free content placement supports an inline-text file or a "
-                "project-contained file/directory source with a destination path.",
-            )
-        ]
-    return result
